chore: update sdk-react-native and sdk-web to use shared package, refactor OrgaAIProvider, and clean up unused files
