+++ .npmrc
@@ -1,0 +1,3 @@
+@orga-ai:registry=https://registry.npmjs.org/
+//registry.npmjs.org/:_authToken=${NPM_TOKEN}
+
