+++ packages/sdk-react-native/package.json
@@ -1,7 +1,10 @@
 {
-  "name": "orga-ai-sdk/react-native",
+  "name": "@orga-ai/sdk-react-native",
   "version": "1.0.0",
   "main": "lib/index.js",
+  "publishConfig": {
+    "access": "restricted"
+  },
   "scripts": {
     "build": "tsc --project tsconfig.json",
     "clean": "rm -rf dist lib build *.tsbuildinfo",
@@ -33,6 +36,6 @@
   },
   "dependencies": {
     "axios": "^1.10.0",
-    "@orga-ai-sdk/shared": "1.0.0"
+    "@orga-ai-sdk/shared": "workspace:*"
   }
 }
