+++ packages/sdk-react-native/src/hooks/useOrgaAI.ts
@@ -8,31 +8,25 @@ import {
 } from "react-native-webrtc";
 import RTCDataChannel from "react-native-webrtc/lib/typescript/RTCDataChannel";
 import InCallManager from "react-native-incall-manager";
-import React, { createContext, useContext } from "react";
-
-import {
-  OrgaAIHookReturn,
-  OrgaAIHookCallbacks,
-  SessionConfig,
-  Transcription,
-  ConnectionState,
-  CameraPosition,
-  IceCandidateEvent,
-} from "../types";
-import {
-  PermissionError,
+import { SessionConfig, OrgaAI,  PermissionError,
   ConnectionError,
   SessionError,
   ConfigurationError,
-} from "../errors";
-import { OrgaAI } from "../core/OrgaAI";
-import {
   getMediaConstraints,
   logger,
   connectToRealtime,
   RTCIceServer,
   RTCIceCandidateInit,
-} from "../utils";
+ } from "@orga-ai-sdk/shared";
+
+import {
+  OrgaAIHookReturn,
+  OrgaAIHookCallbacks,
+  Transcription,
+  ConnectionState,
+  CameraPosition,
+  IceCandidateEvent,
+} from "../types";
 
 
 // Rename the original hook for internal use
