+++ packages/sdk-web/package.json
@@ -31,6 +31,6 @@
   },
   "dependencies": {
     "axios": "^1.10.0",
-    "@orga-ai-sdk/shared": "1.0.0"
+    "@orga-ai-sdk/shared": "workspace:*"
   }
 }
