+++ packages/shared/package.json
@@ -14,9 +14,14 @@
   "dependencies": {
     "axios": "^1.6.0"
   },
+  "peerDependencies": {
+    "react": "^18.0.0"
+  },
   "devDependencies": {
-    "typescript": "^5.0.0",
+    "@types/react": "^19.1.8",
     "jest": "^29.0.0",
-    "ts-jest": "^29.0.0"
+    "react": "^18.0.0",
+    "ts-jest": "^29.0.0",
+    "typescript": "^5.0.0"
   }
 }
