+++ packages/shared/src/index.ts
@@ -1,0 +1,5 @@
+export { OrgaAI } from "./core/OrgaAI";
+export * from "./errors";
+export * from "./hooks/OrgaAIProvider";
+export * from "./types";
+export * from "./utils";
