+++ packages/shared/tsconfig.json
@@ -1,0 +1,21 @@
+{
+    "compilerOptions": {
+      "target": "es2020",
+      "lib": ["dom", "esnext"],
+      "module": "commonjs",
+      "declaration": true,
+      "outDir": "./lib",
+      "rootDir": "./src",
+      "strict": true,
+      "esModuleInterop": true,
+      "skipLibCheck": true,
+      "forceConsistentCasingInFileNames": true,
+      "moduleResolution": "node",
+      "resolveJsonModule": true,
+      "isolatedModules": true,
+      "noEmit": false,
+      "jsx": "react"
+    },
+    "include": ["src/**/*"],
+    "exclude": ["node_modules", "lib", "**/*.test.ts", "**/*.test.tsx"]
+  }
