--- project-tree.txt
@@ -1,69 +1,0 @@
-./
-├── .gitignore
-├── .npmrc
-├── package.json
-├── packages/
-│   ├── eslint-config/
-│   │   ├── base.js
-│   │   ├── next.js
-│   │   ├── package.json
-│   │   ├── react-internal.js
-│   │   └── README.md
-│   ├── sdk-react-native/
-│   │   ├── package.json
-│   │   ├── src/
-│   │   │   ├── __tests__/
-│   │   │   │   ├── OrgaAI.test.ts
-│   │   │   │   ├── useOrgaAi.test.ts
-│   │   │   │   └── utils.test.ts
-│   │   │   ├── components/
-│   │   │   │   └── OrgaAICameraView.tsx
-│   │   │   ├── core/
-│   │   │   │   └── OrgaAI.ts
-│   │   │   ├── errors/
-│   │   │   │   └── index.ts
-│   │   │   ├── hooks/
-│   │   │   │   ├── OrgaAIProvider.tsx
-│   │   │   │   └── useOrgaAI.ts
-│   │   │   ├── index.ts
-│   │   │   ├── types/
-│   │   │   │   └── index.ts
-│   │   │   └── utils/
-│   │   │       └── index.ts
-│   │   └── tsconfig.json
-│   ├── sdk-web/
-│   │   ├── package.json
-│   │   ├── src/
-│   │   │   ├── __tests__/
-│   │   │   ├── components/
-│   │   │   │   ├── OrgaAudio.tsx
-│   │   │   │   └── OrgaVideo.tsx
-│   │   │   ├── core/
-│   │   │   │   └── OrgaAI.ts
-│   │   │   ├── errors/
-│   │   │   │   └── index.ts
-│   │   │   ├── hooks/
-│   │   │   │   ├── OrgaAIProvider.tsx
-│   │   │   │   └── useOrgaAI.ts
-│   │   │   ├── index.ts
-│   │   │   ├── types/
-│   │   │   │   └── index.ts
-│   │   │   └── utils/
-│   │   │       └── index.ts
-│   │   └── tsconfig.json
-│   ├── shared/
-│   │   ├── package.json
-│   │   └── src/
-│   │       └── api.ts
-│   └── typescript-config/
-│       ├── base.json
-│       ├── nextjs.json
-│       ├── package.json
-│       └── react-library.json
-├── pnpm-lock.yaml
-├── pnpm-workspace.yaml
-├── project-tree.txt
-├── README.md
-└── turbo.json
-
-24 directories, 43 files
